requirements.txt
numpy==1.19.3
opencv_python==4.5.3.56
Pillow==9.5.0
streamlit==1.20.0
streamlit_webrtc==0.45.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
